+++ 5383/maksimova/5/Square.cpp
@@ -18,7 +18,7 @@ bool Square::is_square(const std::vector<Dot>& v) {
 
 void Square::print(std::ostream& OS) const {
 	OS << "SQUARE" << std::endl;
-	OS << "Color: " << color.R << " " << color.G << " " << color.B << std::endl;
+	OS << "Color: " <<(int) color.R << " " << (int) color.G << " " << (int)color.B << std::endl;
 	OS << "id: " << id << std::endl;
 }
